Project API

See merge request projects-dashboard/projects-dashboard-api!2

+++ src/models/developer/developer.service.ts
@@ -1,6 +1,8 @@
 import * as createHttpError from "http-errors";
 
-import { SuccessResponse } from "../../types/common-types";
+import { SuccessResponse } from "../../common/types/common-types";
+
+import { ProjectsService } from "../project/projects.service";
 
 import { DeveloperRepository } from "./repositories/developer.repository";
 import { Developer } from "./entities/developer.entity";
@@ -10,15 +12,19 @@ import { DeveloperSerializer } from "./serializers/developer.serializer";
 export class DeveloperService {
   private developerRepository = new DeveloperRepository(Developer);
 
-  async create(next, entity: DeveloperDTO): Promise<DeveloperSerializer> {
-    const developerFromDB = await this.developerRepository.get(next, { email: entity.email });
+  private projectService = new ProjectsService();
+
+  async create(next, developer: DeveloperDTO): Promise<{ developer: DeveloperSerializer }> {
+    const developerFromDB = await this.developerRepository.get(next, { email: developer.email });
     if (developerFromDB) {
       return next(
         new createHttpError.BadRequest(`Developer with this email address already exists.`)
       );
     }
 
-    return await this.developerRepository.createEntity(next, entity);
+    const newDeveloper = await this.developerRepository.createEntity(next, developer);
+
+    return { developer: newDeveloper };
   }
 
   async getAll(next): Promise<{ developers: DeveloperSerializer[] }> {
@@ -31,6 +37,13 @@ export class DeveloperService {
   }
 
   async delete(next, uuid: string): Promise<SuccessResponse> {
+    const developer = await this.developerRepository.get(next, { uuid }, ["features"], true, true);
+    if (!developer) {
+      return;
+    }
+
+    await this.projectService.unassignDeveloperFromFeatures(next, developer.features);
+
     return await this.developerRepository.deleteEntity(next, { uuid });
   }
 }
